+++ src/pl/edu/pw/elka/prm2t/lab3/zad1/ProduktWKoszyku.java
@@ -3,7 +3,7 @@ package pl.edu.pw.elka.prm2t.lab3.zad1;
 public class ProduktWKoszyku extends Produkt {
     private int liczbaSztukProduktu;
 
-    public ProduktWKoszyku(String nazwa, float cena, int liczbaSztukProduktu) {
+    public ProduktWKoszyku(String nazwa, double cena, int liczbaSztukProduktu) {
         super(nazwa, cena);
         this.liczbaSztukProduktu = liczbaSztukProduktu;
     }
